Add back window config mutation when setting modes
This doesn't really do anything, but we should probably have them there just to prevent some
obscure bug in the future (or weird use case).

NOTE: it is sort of unclear from the GDX specification and/or from other backend implementations
whether or not the application configuration is actually intended to be dynamically modified as
settings are updated. It seems like it should be a one-time initial configuration kind of thing,
but who the hell really knows...