Add setUndeocrated/setResizable APIs to Graphics.
Add implemenations for all backends.

Note: most desktop backend implementations require a call to 'setWindowedMode' in order for the changes to take effect.

It is worth noting that this is not necessary with the JogAmp backend. However, as of the time of writing, that backend is not included in LibGDX master.

Fixes #3773